requirements.txt
Django>=4.2
djangorestframework>=3.14.0
pillow>=9.5.0
django-environ>=0.10.0
